+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+## [1.3.1] - 24/06/2021
 
-* TODO: Describe initial release.
+Fixed:
+- An issue with too many workers being created and not being cleared from the Watcher.
+
+## [1.3.0] - 23/06/2021
+
+Initial release! Adds Flutter specific Widgets to use with Styx to inject entities into your widget
+tree. See example for more info.